SMILE-775-thumbnailer-my-resources-user-test-feedback, functionality implemented
